lab4完成

+++ src/php/collection_delete.php
@@ -13,12 +13,12 @@
 <?php
 @$list_id = $_POST["listID"];
 @$show_id = $_POST["showID"];
+@$user = $_POST["userID"];
 echo "<script>console.log('$list_id');</script>";
 if ($list_id) {
    echo "<script>clearVal('$show_id');</script>";
-    $sql = "DELETE FROM `wishlist` WHERE `wishlist`.`listID` = '$list_id'";
+    $sql = "DELETE FROM `wishlist` WHERE `wishlist`.`listID` = '$list_id' AND userID = $user";
    if ($conn->query($sql) === TRUE) {
        //echo '<script>alert(\'删除成功\')</script>';
-
    }
 }
